Add UI.png
Add UI.png to creates the product website.
And edit some mistakes regarding naming and code quality.

Add a screenshot of the program as the product website.

To make sure the potential user able to see the GUI.
