--- src/main/java/duke/Launcher.java
@@ -1,14 +1,0 @@
-package duke;
-
-import javafx.application.Application;
-
-/**
- * Launcher code.
- *
- *
- */
-public class Launcher {
-    public static void main(String[] args) {
-        Application.launch(Main.class, args);
-    }
-}
