--- src/main/java/duke/command/Commands.java
@@ -1,9 +1,0 @@
-package duke.command;
-
-/**
- * Enum for all valid commands.
- *
- */
-public enum Commands {
-    LIST, TODO, EVENT, DEADLINE, MARK, UNMARK, DELETE, BYE, FIND;
-}
