--- src/main/java/duke/exception/CommandFormatException.java
@@ -1,18 +1,0 @@
-package duke.exception;
-
-
-/**
- * Exception used when command format is wrong.
- *
- */
-public class CommandFormatException extends Exception {
-
-    /**
-     * Constructor for CommandFormatException.
-     *
-     */
-    public CommandFormatException() {
-        super();
-    }
-}
-
