--- src/main/java/duke/exception/CommandInvalidException.java
@@ -1,16 +1,0 @@
-package duke.exception;
-
-/**
- * Exception used when command is invalid.
- *
- */
-public class CommandInvalidException extends Exception {
-
-    /**
-     * Constructor for CommandInvalidException.
-     *
-     */
-    public CommandInvalidException(String e) {
-        super(e);
-    }
-}
