--- src/main/java/duke/exception/DescriptionFormatException.java
@@ -1,17 +1,0 @@
-package duke.exception;
-
-/**
- * Exception used when description format is wrong.
- *
- */
-public class DescriptionFormatException extends Exception {
-
-    /**
-     * Constructor for DescriptionFormatException.
-     *
-     * @param e error message.
-     */
-    public DescriptionFormatException(String e) {
-        super(e);
-    }
-}
