+++ src/main/java/ganAnWo/Launcher.java
@@ -1,0 +1,14 @@
+package ganAnWo;
+
+import javafx.application.Application;
+
+/**
+ * Launcher code.
+ *
+ *
+ */
+public class Launcher {
+    public static void main(String[] args) {
+        Application.launch(Main.class, args);
+    }
+}
