+++ src/main/java/ganAnWo/command/Commands.java
@@ -1,0 +1,9 @@
+package ganAnWo.command;
+
+/**
+ * Enum for all valid commands.
+ *
+ */
+public enum Commands {
+    LIST, TODO, EVENT, DEADLINE, MARK, UNMARK, DELETE, BYE, FIND;
+}
