+++ src/main/java/ganAnWo/exception/CommandFormatException.java
@@ -1,0 +1,18 @@
+package ganAnWo.exception;
+
+
+/**
+ * Exception used when command format is wrong.
+ *
+ */
+public class CommandFormatException extends Exception {
+
+    /**
+     * Constructor for CommandFormatException.
+     *
+     */
+    public CommandFormatException() {
+        super();
+    }
+}
+
