+++ src/main/java/ganAnWo/exception/CommandInvalidException.java
@@ -1,0 +1,16 @@
+package ganAnWo.exception;
+
+/**
+ * Exception used when command is invalid.
+ *
+ */
+public class CommandInvalidException extends Exception {
+
+    /**
+     * Constructor for CommandInvalidException.
+     *
+     */
+    public CommandInvalidException(String e) {
+        super(e);
+    }
+}
